migrated an existing  application to  MVC pattern format
