+++ android/local.properties
@@ -1,5 +1,5 @@
-sdk.dir=E:\\Users\\devke\\AppData\\Local\\Android\\Sdk
-flutter.sdk=E:\\path-apps\\flutter
+sdk.dir=C:\\Users\\irfan\\AppData\\Local\\Android\\sdk
+flutter.sdk=C:\\src\\flutter
 flutter.buildMode=debug
 flutter.versionName=1.0.0
 flutter.versionCode=1
