+++ ios/Flutter/Generated.xcconfig
@@ -1,6 +1,6 @@
 // This is a generated file; do not edit or check into version control.
-FLUTTER_ROOT=E:\path-apps\flutter
-FLUTTER_APPLICATION_PATH=F:\Users\devke\Documents\group-projects\hacktoberfest-flutter\hacktoberfest_flutter
+FLUTTER_ROOT=C:\src\flutter
+FLUTTER_APPLICATION_PATH=C:\Users\irfan\Desktop\habeeb\hacktoberfest-flutter23
 COCOAPODS_PARALLEL_CODE_SIGN=true
 FLUTTER_TARGET=lib\main.dart
 FLUTTER_BUILD_DIR=build
