+++ lib/helpers/colors.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+Color hacktoberBlue = const Color(0xFF92c1da);
+
+Color githubButtonGreen = const Color(0xff2EA44F);
+
+Color whiteColor = const Color(0xffFFFFF4);
+
+Color hacktoberPink = const Color(0xFFff8ae0);
+
+Color hacktoberPurple = const Color(0xFF9b4667);
+
+Color hactoberViolet = const Color.fromRGBO(153, 69, 255, 1);
+
+Color darkModeHactober = const Color(0xff183D5D);
+
+
